Quiet unnecessary warnings
To avoid warnings, both values in CHECK_xx() must be the same type.

+++ omx/SoftFFmpegAudio.cpp
@@ -1108,7 +1108,7 @@ int32_t SoftFFmpegAudio::decodeAudio() {
     int len = 0;
     int gotFrm = false;
     int32_t ret = ERR_OK;
-    int32_t inputBufferUsedLength = 0;
+    uint32_t inputBufferUsedLength = 0;
     bool is_flush = (mEOSStatus == OUTPUT_FRAMES_FLUSHED);
     List<BufferInfo *> &inQueue = getPortQueue(kInputPortIndex);
     BufferInfo *inInfo = NULL;
@@ -1404,7 +1404,7 @@ void SoftFFmpegAudio::drainAllOutputBuffers() {
                 drainEOSOutputBuffer();
                 return;
             } else {
-                CHECK_EQ(err, ERR_OK);
+                CHECK_EQ(err, (int32_t)ERR_OK);
             }
         }
 
@@ -1473,7 +1473,7 @@ void SoftFFmpegAudio::onQueueFilled(OMX_U32 /* portIndex */) {
                 CHECK_EQ(mResampledDataSize, 0);
                 continue;
             } else {
-                CHECK_EQ(err, ERR_OK);
+                CHECK_EQ(err, (int32_t)ERR_OK);
             }
         }
